+++ example_service.go
@@ -1,9 +1,8 @@
-package example
+package tracr
 
 import (
 	"context"
 	"fmt"
-	"github.com/micklove/tracr/internal/tracr"
 )
 
 // Example service... For each func, the correlation id will be present in the context, since it was added to the context by the middleware
@@ -13,7 +12,7 @@ type Service struct {
 
 // Echo return the correlation id from the given context
 func (s Service) Echo(ctx context.Context) (string, error) {
-	correlationID, err := tracr.GetCID(ctx)
+	correlationID, err := GetCID(ctx)
 	if err != nil {
 		return "", err
 	}
